% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{check_frame_type}
\alias{check_frame_type}
\title{Check if the frame is a valid data structure}
\usage{
check_frame_type(frame)
}
\arguments{
\item{frame}{An object to be validated as a sampling frame.}
}
\value{
Stops execution if validation fails. Invisibly returns TRUE.
}
\description{
Ensures that the input frame is a data.frame, data.table, or tibble.
}
\keyword{internal}
