% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{check_n}
\alias{check_n}
\title{Validate the sample size \code{n}}
\usage{
check_n(n, frame, curstrat, n_le_N = FALSE)
}
\arguments{
\item{n}{Desired sample size.}

\item{frame}{Data frame or data table of the sampling frame.}

\item{curstrat}{Optional character vector indicating the stratum.}

\item{n_le_N}{Logical. Should \code{n} be forced to be less than or equal to \code{N}?}
}
\value{
Stops with an error or outputs a message if \code{n == N}. Returns nothing.
}
\description{
Performs a set of checks on the sample size, including numeric type, bounds, and optional strata.
}
\keyword{internal}
