% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{check_outall}
\alias{check_outall}
\title{Validate the 'outall' argument}
\usage{
check_outall(outall)
}
\arguments{
\item{outall}{A logical value indicating whether to return the full frame or only selected rows.}
}
\value{
Stops execution if validation fails. Invisibly returns TRUE.
}
\description{
Checks that the outall argument is a single logical value.
}
\keyword{internal}
