% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chromy_pps.R
\name{chromy_inner}
\alias{chromy_inner}
\title{Select a sequential PPS sample given expected number of hits}
\usage{
chromy_inner(exphits)
}
\arguments{
\item{exphits}{A vector of non-negative values}
}
\value{
Returns a vector the same length of exphits with the number of random hits for each unit
}
\description{
Select a sequential PPS sample given expected number of hits
}
\references{
Chromy, J. R. (1979). “Sequential Sample Selection Methods.”
In \emph{Proceedings of the Survey Research Methods Section}, 401–406.
Washington, DC: American Statistical Association.
\url{http://www.asasrms.org/Proceedings/papers/1979_081.pdf}
}
\keyword{internal}
