% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chromy_pps.R
\name{chromy_pps}
\alias{chromy_pps}
\alias{seq_pps}
\title{Select a sequential PPS sample}
\usage{
chromy_pps(frame, n, mos, outall = FALSE, curstrat = NULL)
}
\arguments{
\item{frame}{The input data frame for the function to work on.}

\item{n}{The sample size, the parameter expects an integer of length 1.  The function will check if n is less than or equal to the number of rows in the input frame.}

\item{mos}{The measure of size, the parameter expects a character string to indicate the variable to be used as the measure of size. The variable must exist on the frame and be a non-missing and non-negative numeric variable.}

\item{outall}{Output all records or selected records.  If outall is TRUE, then all records are returned and the following variables are created: SelectionIndicator, SamplingWeight, NumberHits, and ExpectedHits.  If outall is FALSE, then the selected records are returned and the following variables are created:  SamplingWeight, NumberHits, ExpectedHits.}

\item{curstrat}{A character variable that specifies the current strata, only used as an assertion for the n == N test.}
}
\value{
Returns an object of type tidytable that contains
the weight, expected hits (selection probability for nonreplacement designs), and
number of hits plus all original variables.
Include a SelectionIndicator variable if outall=TRUE
}
\description{
Draws a sequential sample of size n.
Each unit's probability of selection is proportional to its size measure.
This is a minimum replacement method as discussed in Chromy (1979).
}
\examples{
# PPS sample of counties using population size as MOS
# LA county will be selected two or three times based on expected hits
# Cook, Harris, and Maricopa will be selected one or two times based on expected hits

county_2023 |>
  tidytable::select(GEOID, Name, Pop_Tot) |>
  chromy_pps(n = 75, mos = "Pop_Tot") |>
  tidytable::arrange(desc(ExpectedHits))

county_2023 |>
  tidytable::select(GEOID, Name, Pop_Tot) |>
  chromy_pps(n = 75, mos = "Pop_Tot", outall = TRUE) |>
  tidytable::arrange(desc(ExpectedHits))

}
\references{
Chromy, J. R. (1979). “Sequential Sample Selection Methods.”
In \emph{Proceedings of the Survey Research Methods Section}, 401–406.
Washington, DC: American Statistical Association.
\url{http://www.asasrms.org/Proceedings/papers/1979_081.pdf}
}
