% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_group_variable.R
\name{create_group_variable}
\alias{create_group_variable}
\title{Create grouping variable}
\usage{
create_group_variable(.data, groupvars, groupid)
}
\arguments{
\item{.data}{A data frame, tibble, or data.table. No default.}

\item{groupvars}{A vector of character variables which are the names of the columns to sort on. No default.}

\item{groupid}{ID grouping variable name. No default.}
}
\value{
A copy of the input data with the addition of the grouping ID variable
}
\description{
Given a data frame and a vector of variables, creates a new variable that assigns a unique ID to each group of rows that have the same values
for the given variables. The function will return the data frame with the new id variable added.
}
\keyword{internal}
