% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/placeholder.R
\name{placeholder}
\alias{placeholder}
\title{Placeholder function}
\usage{
placeholder(input)
}
\arguments{
\item{input}{some sort of input}
}
\value{
always returns the value a data frame with value x 1 through 9
}
\description{
Evaluate if functions are being processed correctly by R CMD CHECK.
}
