% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/placeholder.R
\name{return_0}
\alias{return_0}
\title{Placeholder function}
\usage{
return_0(input)
}
\arguments{
\item{input}{some sort of input}
}
\value{
always returns the value 0
}
\description{
Evaluate if functions are being processed correctly by R CMD CHECK.
}
