% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serp_sort.R
\name{serp_sort}
\alias{serp_sort}
\title{Serpentine sort function}
\usage{
serp_sort(.data, ..., naorder = TRUE, sortID = FALSE)
}
\arguments{
\item{.data}{A data frame. No default.}

\item{...}{Variables to sort by - must be listed as character variables. No default.}

\item{naorder}{Logical value for whether to put NA values at the end of the sorted data frame. Default value is TRUE and treats
NA values as the largest values when sorting. A value of FALSE treats NA Values as the smallest values when sorting.}

\item{sortID}{Logical value that indicates whether the sort group information used to implement the sorting should be added to .data.
By default, sort group is FALSE which means the sort group information is not added to .data. When TRUE, and there are
k variables to sort by, a set of k-1 sort group variables with names sortSerpj (j=1 to k-1) are added to .data.}
}
\value{
A copy of .data that is hierarchically sorted and, if specified, includes the sort group variables created for sorting
}
\description{
Sorts the rows in a given data frame in a hierarchical fashion given a list of variables (...)
}
\examples{
# Sort counties by Region, then Division, in a serpentine pattern
county_2023 |>
  serp_sort("Region", "Division")

# Keep the intermediate sort-group columns (sortSerp1, sortSerp2)
county_2023 |>
  serp_sort("Region", "Division", sortID = TRUE)

}
