% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srs.R
\name{srs}
\alias{srs}
\title{Simple random sampling function}
\usage{
srs(frame, n, outall = FALSE, curstrat = NULL)
}
\arguments{
\item{frame}{A data frame, data.table, or tibble from which to draw the sample. No default.}

\item{n}{The sample size to draw from the frame. Must be a positive integer. No default.}

\item{outall}{A logical value indicating whether to return the entire frame with a selection indicator or just the sample. Default is FALSE.}

\item{curstrat}{A character string indicating the current stratum. Used only for printing messages when n = N. Default is NULL.}
}
\value{
A tidytable object containing the entire frame with a selection indicator or just the sample, dependent on the value of outall.
Selection probability and sampling weight are also included. The sample size, n, and the population size, N, are printed to the console.
}
\description{
Draws a simple random sample of size n.
}
\examples{
#Random sample of 200 universities, only sampled rows returned
ipeds |>
  tidytable::filter(!is.na(ENRTOT)) |>
  srs(n = 200, outall = FALSE)

# Return full dataset with selection indicators
ipeds |>
  tidytable::filter(!is.na(ENRTOT)) |>
  srs(n = 200, outall = TRUE)



}
