% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sys.R
\name{sys}
\alias{sys}
\title{Systematic Sampling Without Replacement}
\usage{
sys(frame, n, curstrat = NULL, outall = FALSE)
}
\arguments{
\item{frame}{a \code{data.frame}, \code{tibble}, or \code{data.table} containing the sampling frame. Must have at least one row.}

\item{n}{Integer. The desired sample size. Must be less than or equal to the number of rows in \code{frame}.}

\item{curstrat}{Character or NULL. Optional stratum name for printing messages.}

\item{outall}{logical indicator for whether full frame is returned or just the sample}
}
\value{
a \code{data.table} with the original columns plus:
\describe{
\item{SelectionProbability}{Equal to n / N for all units.}
\item{SamplingWeight}{Equal to N / n for all units.}
\item{SelectionIndicator}{TRUE if selected, FALSE otherwise.-only included if \code{outall=TRUE}}
\item{NumberHits}{1 if selected, 0 otherwise.}
\item{ExpectedHits}{Equal to SelectionProbability.}
}
}
\description{
Draws a systematic sample of size \code{n} from a data frame. Each unit has an equal probability of being selected.
}
\examples{

#Sort by REGION, DIVISION, and Pop_Tot, then take a sample
puma_2023 |>
  tidytable::arrange(Region, Division, Pop_Tot) |>
  sys(n = 50, outall = FALSE)

#Return full dataset with selection indicators
puma_2023 |>
  tidytable::arrange(Region, Division, Pop_Tot) |>
  sys(n = 50, outall = TRUE)



}
\references{
Kalton, G. (1983). \emph{Introduction to Survey Sampling}. SAGE Publications. https://doi.org/10.4135/9781412984683
}
