% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allocate.R
\name{allocate}
\alias{allocate}
\title{Sample allocation}
\usage{
allocate(
  allocation,
  N.h,
  n.samp = NULL,
  S.h = NULL,
  c.h = NULL,
  cost = NULL,
  variance = NULL,
  power = NULL,
  lbound = 2
)
}
\arguments{
\item{allocation}{type of allocation,
must be one of \code{"proportional"}, \code{"power"}, \code{"neyman"}, or \code{"optimal"}.}

\item{N.h}{vector of population stratum sizes (\eqn{N_h}, all positive values), for example \code{c(150, 600, 250)}. \cr\cr
required for all allocation types.}

\item{n.samp}{total sample size to be allocated (positive integer of length 1). \cr\cr
required for the following allocation types: proportional, power, and Neyman, and \code{NULL} otherwise.}

\item{S.h}{vector of stratum unit standard deviations
(positive values same length as \code{N.h}) (\eqn{S_h}). \cr\cr
required for the following allocation types: Neyman, and optimal, and \code{NULL} otherwise.}

\item{c.h}{vector of cost per unit in stratum h (positive values same length as \code{N.h}) (\eqn{c_h}). \cr\cr
required for the optimal allocation only, and \code{NULL} otherwise.}

\item{cost}{total variable cost (positive value) \eqn{(C – c_0)}. \cr\cr
required for the cost-constrained optimal
allocation only, and \code{NULL} otherwise.}

\item{variance}{fixed variance target for estimated mean (positive value) (\eqn{V_0}). \cr\cr
required for the precision-constrained optimal allocation only, and \code{NULL} otherwise.}

\item{power}{power value for power allocation (\eqn{0 \le \alpha \le 1}). \cr\cr
required for the power allocation only, and \code{NULL} otherwise.}

\item{lbound}{minimum stratum-level
(positive integer of length 1). Default value is 2.}
}
\value{
Integer vector same length of N.h final allocation
}
\description{
Compute the proportional, power, Neyman, and optimal sample allocations.
}
\section{Method}{

The \emph{allocate} function allocates a sample size \emph{n} on \emph{H} strata using one of the following allocation methods:
\enumerate{
\item Proportional allocation [\code{n.samp, N.h, allocation = "proportional"}]
\deqn{n_h = n \times \frac{N_h}{\sum\limits_{h=1}^H N_h}}
where \cr
\eqn{n}: total sample size to be allocated (function input is \code{n.samp}), and \cr
\eqn{N_h}: population size of stratum \emph{h} (function input is \code{N.h}).
\item Power allocation [\code{n.samp, N.h, power, allocation = "power"}]
\deqn{n_h = n \times \frac{N_h^\alpha}{\sum\limits_{h=1}^H N_h^\alpha}}
where \cr
\eqn{\alpha}: a power value to control over-under-sampling with \eqn{0 \le \alpha \le 1} (function input is \code{power}).
\item Neyman allocation [\code{n.samp, N.h, S.h, allocation = "neyman"}]
\deqn{n_h = n \times \frac{N_h S_h}{\sum\limits_{h=1}^H N_h S_h}}
where \cr
\eqn{S_h}:  standard deviation of stratum \emph{h} (function input is \code{S.h}).
\item Optimal allocation
\itemize{
\item cost-constrained [\code{N.h, S.h, c.h, cost, allocation = "optimal"}]
\deqn{n_h = (C−c_0) \times \frac{N_h S_h / \sqrt{c_h}}{\sum\limits_{h=1}^H N_h S_h \sqrt{c_h}}}
            where \cr
\eqn{c_h}: cost per unit in stratum \emph{h} (function input is \code{c.h}), and \cr
\eqn{(C – c_0)}: total variable cost (function input is \code{cost})
\item precision-constrained [\code{N.h, S.h, c.h, variance, allocation = "optimal"}]
\deqn{n_h = N_h S_h / \sqrt{c_h} \times \frac{\sum\limits_{h=1}^H N_h S_h \sqrt{c_h}}{V_0 \left(\sum\limits_{h=1}^H  N_h \right)^2 + \sum\limits_{h=1}^H N_h S_h^2}}
where \cr
\eqn{V_0}: fixed variance target for estimated mean (function input is \code{variance})
}
}

The table below presents the relevant inputs for each type; when irrelevant inputs are entered, an error message will be displayed.

\tabular{lllll}{
\strong{allocation} \tab \strong{N.h} \tab \strong{n.h} \tab \strong{power} \tab \strong{lbound} \cr
proportional \tab ✓ \tab ✓  \tab   \tab ✓\cr
power        \tab ✓ \tab ✓  \tab ✓ \tab ✓
}
}

\examples{
# Example of proportional allocation
ipeds_summary |>
  tidytable::mutate(
    n = allocate("proportional", N.h = N, n.samp = 500)
  )
# Example of power allocation
ipeds_summary |>
  tidytable::mutate(
    n = allocate("power", N.h = N, power = 0.5, n.samp = 500)
  )
}
