% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chromy_pps.R
\name{chromy_pps}
\alias{chromy_pps}
\title{Select a systematic PPS sample}
\usage{
chromy_pps(frame, n, mos, outall = FALSE, curstrat = NULL)
}
\arguments{
\item{frame}{The input data frame for the function to work on.}

\item{n}{The sample size, the parameter expects an integer of length 1.  The function will check if n is less than or equal to the number of rows in the input frame.}

\item{mos}{The measure of size, the parameter expects a character string to indicate the variable to be use as the measure of size.  The variable must exists on the frame and be non-missing and non-negative numeric variable.}

\item{outall}{Output all records or selected records.  If outall is TRUE, then all records are return and the following variables are created: SelectionIndicator, SamplingWeight, NumberHits, and ExpectedHits.  If outall is FALSE, then the selected records are return and the following variables are created:  SamplingWeight, NumberHits, ExpectedHits.}

\item{curstrat}{A character variable that specifies the current strata, only used as an assertion for the n == N test.}
}
\value{
Returns an object of type tidytable that contains the weight, selection probability, number of hits, etc plus all original variables.
}
\description{
Draws a sequential sample of size n. Each unit’s probability of selection is proportional to its size measure. This is a minimum replacement method.
}
