% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SampleSelectR-package.R
\docType{package}
\name{SampleSelectR-package}
\alias{SampleSelectR}
\alias{SampleSelectR-package}
\title{SampleSelectR: Randomly Select Samples for Various Probability-Based Methods}
\description{
Randomly select samples using simple random sampling (SRS), systematic sampling, and various probability proportional to size (PPS) methods, including systematic PPS and sequential PPS (i.e., Chromy's method). Also includes functionality to allocate sample sizes across strata using proportional, power, Neyman, and optimal allocation methods, and to select samples within strata. Designed to make survey sample design and selection reproducible, efficient, and transparent for survey statisticians and researchers.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/RTIInternational/SampleSelectR}
  \item \url{https://rtiinternational.github.io/SampleSelectR/}
  \item Report bugs at \url{https://github.com/RTIInternational/SampleSelectR/issues}
}

}
\author{
\strong{Maintainer}: Stephanie Zimmer \email{sazimmer@rti.org}

Authors:
\itemize{
  \item John David Bunker, Jr.
  \item Thomas Burkett
  \item Philip Lee
  \item Haby Sow
}

Other contributors:
\itemize{
  \item  RTI International [copyright holder, funder]
}

}
\keyword{internal}
