% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SampleSelectR-package.R
\docType{package}
\name{SampleSelectR-package}
\alias{SampleSelectR}
\alias{SampleSelectR-package}
\title{SampleSelectR: Randomly select samples for various probability-based methods}
\description{
Randomly select samples with SRS, systematic, and various PPS methods. Also includes functionality to select within strata and allocate sample sizes.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/RTIInternational/SampleSelectR}
  \item Report bugs at \url{https://github.com/RTIInternational/SampleSelectR/issues}
}

}
\author{
\strong{Maintainer}: Stephanie Zimmer \email{sazimmer@rti.org}

Authors:
\itemize{
  \item John David Bunker, Jr.
  \item Thomas Burkett
  \item Philip Lee
  \item Haby Sow
}

Other contributors:
\itemize{
  \item  RTI International [funder]
}

}
\keyword{internal}
