% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SampleSelectR-package.R
\docType{package}
\name{SampleSelectR-package}
\alias{SampleSelectR}
\alias{SampleSelectR-package}
\title{SampleSelectR: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com}

}
\keyword{internal}
