% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_sample.R
\name{select_sample}
\alias{select_sample}
\title{Sample selection function}
\usage{
select_sample(
  frame,
  method,
  n,
  outall = FALSE,
  strata = NULL,
  mos = NULL,
  sort_vars = NULL,
  sort_method = NULL
)
}
\arguments{
\item{frame}{A data frame, data.table, or tibble from which to draw the sample. No default.}

\item{method}{The desired sampling method. Valid options are "srs", "sys_eq", "sys_pps", and "chromy_pps". No default.}

\item{n}{The sample size to draw from the frame. If strata is NULL, must be a positive integer. If strata is not NULL, must be a
data.frame, tibble, or data.table with columns for each stratification variable as the same type and variable names as the frame
plus a column with the sample size (\code{sample_size}) which is a positive integer. No default.}

\item{outall}{A logical value indicating whether to return the entire frame with a selection indicator or just the sample. Default is FALSE.}

\item{strata}{A vector of characters with variable names of strata. Default is NULL.}

\item{mos}{A character string defining the variable name on the frame for the measure of size. If not NULL, must have method = c("sys_pps", "chromy_pps").
If NULL, must have method=c("srs", "sys"). Default is NULL.}

\item{sort_vars}{A character string indicating the variables that should be used to sort the frame. If not NULL, cannot have method = "srs".
Default is NULL.}

\item{sort_method}{A character string defining the method to implicitly sort the frame. Valid options are "serpentine" and "nest". Must coincide
with sort_var; i.e., both must be NULL or both must be not NULL. Default is NULL.}
}
\value{
A tidytable object containing the entire frame with a selection indicator or just the sample, dependent on the value of outall.
Selection probability and sampling weight are also included. May include various summary messages to the console when applicable for
certain sampling methods.
}
\description{
Selects a random sample using a specified method and sample size. Selection can also optionally be stratified and/or include a measure of
size (mos) if a PPS method is used.
}
